+++ pkg/argsbuilder/update_serving_custom.go
@@ -1,0 +1,89 @@
+// Copyright 2018 The Kubeflow Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License
+package argsbuilder
+
+import (
+	"fmt"
+	"reflect"
+	"strings"
+
+	"github.com/kubeflow/arena/pkg/apis/types"
+	"github.com/spf13/cobra"
+)
+
+type UpdateCustomServingArgsBuilder struct {
+	args        *types.UpdateCustomServingArgs
+	argValues   map[string]interface{}
+	subBuilders map[string]ArgsBuilder
+}
+
+func NewUpdateCustomServingArgsBuilder(args *types.UpdateCustomServingArgs) ArgsBuilder {
+	args.Type = types.CustomServingJob
+	s := &UpdateCustomServingArgsBuilder{
+		args:        args,
+		argValues:   map[string]interface{}{},
+		subBuilders: map[string]ArgsBuilder{},
+	}
+	s.AddSubBuilder(
+		NewUpdateServingArgsBuilder(&s.args.CommonUpdateServingArgs),
+	)
+	s.AddArgValue("default-image", DefaultTfServingImage)
+	return s
+}
+
+func (s *UpdateCustomServingArgsBuilder) GetName() string {
+	items := strings.Split(fmt.Sprintf("%v", reflect.TypeOf(*s)), ".")
+	return items[len(items)-1]
+}
+
+func (s *UpdateCustomServingArgsBuilder) AddSubBuilder(builders ...ArgsBuilder) ArgsBuilder {
+	for _, b := range builders {
+		s.subBuilders[b.GetName()] = b
+	}
+	return s
+}
+
+func (s *UpdateCustomServingArgsBuilder) AddArgValue(key string, value interface{}) ArgsBuilder {
+	for name := range s.subBuilders {
+		s.subBuilders[name].AddArgValue(key, value)
+	}
+	s.argValues[key] = value
+	return s
+}
+
+func (s *UpdateCustomServingArgsBuilder) AddCommandFlags(command *cobra.Command) {
+	for name := range s.subBuilders {
+		s.subBuilders[name].AddCommandFlags(command)
+	}
+}
+
+func (s *UpdateCustomServingArgsBuilder) PreBuild() error {
+	for name := range s.subBuilders {
+		if err := s.subBuilders[name].PreBuild(); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func (s *UpdateCustomServingArgsBuilder) Build() error {
+	for name := range s.subBuilders {
+		if err := s.subBuilders[name].Build(); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}