+++ charts/tfserving/CHANGELOG.md
@@ -25,3 +25,9 @@
 ### 0.6.0
 
 * fix no write permission of /tmp directory
+
+### 0.7.0
+
+* support --config-file
+* support --shell
+* support arena serve update tensorflow