+++ charts/triton/CHANGELOG.md
@@ -16,3 +16,9 @@
 ### 0.4.0
 
 * support custom command
+
+### 0.5.0
+
+* support --config-file
+* support --shell
+* support arena serve update triton